+++ ChoGGi's Library/Code/Classes_UI.lua
@@ -57,12 +57,15 @@ local text_style2 = "ChoGGi_TextList12"
 local text_style3 = "ChoGGi_Text14"
 local dlg_border_width = 2
 if ChoGGi.testing then
---~ 	text_style1 = "ChoGGi_Text14"
---~ 	text_style2 = "ChoGGi_TextList14"
---~ 	text_style3 = "ChoGGi_Text16"
-	text_style1 = "ChoGGi_Text16"
-	text_style2 = "ChoGGi_TextList16"
-	text_style3 = "ChoGGi_Text18"
+	if terminal.desktop.box:sizex() > 1920 then
+		text_style1 = "ChoGGi_Text14"
+		text_style2 = "ChoGGi_TextList14"
+		text_style3 = "ChoGGi_Text16"
+	else
+		text_style1 = "ChoGGi_Text16"
+		text_style2 = "ChoGGi_TextList16"
+		text_style3 = "ChoGGi_Text18"
+	end
 	dlg_border_width = 3
 end